% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{animate_rosettes}
\alias{animate_rosettes}
\title{Create an animation of radiating rosettes}
\usage{
animate_rosettes(
  n_layers,
  output_path = getwd(),
  filename = "animation.svg",
  durations = c(draw = 2, hold = 0.5, fade = 0.5, blank = 0.25),
  scale = 1.4,
  tree_color = "#f87580",
  background_color = "#a5e8e3",
  max_rosettes = 5,
  seed_number = 69
)
}
\arguments{
\item{n_layers}{Numeric. The number of layers of nodes. Must be greater than 3, should probably be less than 8.}

\item{output_path}{Character string. The filepath that the result should be written to. Defaults to the current working directory.}

\item{filename}{Character string. The filename for the output. Defaults to "animation.svg".}

\item{durations}{Named numeric vector. The duration in seconds that each step of the animation should take per rosette. Must include values named "draw", "hold", "fade", and "blank". Defaults to \code{c(draw = 2, hold = 0.5, fade = 0.5, blank = 0.25)}.}

\item{scale}{Numeric. The relative scaling to apply to the rosette spacing. Larger values will place the trees farther out from the center. Defaults to \code{1.4}.}

\item{tree_color}{Character string. The hex code for the color to draw the trees in. Defaults to \code{"#f87580"}.}

\item{max_rosettes}{Numeric. The maximum number of rosettes to include in the animation. Defaults to \code{5}.}

\item{seed_number}{Numeric. The seed number for the randomized processes. Defaults to \code{69}.}

\item{background_Color}{Character string.  The hex code for the color to draw the background in. Defaults to \code{"#a5e8e3"}.}
}
\description{
Creates a looping animation of rosettes radiating out from the center and fading, one at a time.
}
