% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{arrange_rosettes}
\alias{arrange_rosettes}
\title{Arrange binary trees into hexagonal rosettes.}
\usage{
arrange_rosettes(trees, scale = 1.4, seed_number = 69)
}
\arguments{
\item{trees}{Data frame. Output from \code{generate_trees()}.}

\item{scale}{Numeric. The relative scaling to apply to the rosette spacing. Larger values will place the trees farther out from the center. Defaults to \code{1.4}.}

\item{seed_number}{Numeric. The seed number for the randomization of assignment of trees to rosettes. Defaults to \code{69}.}
}
\description{
Group trees into hexagons, rotating and translating them appropriately.
}
