% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_trees}
\alias{generate_trees}
\title{Generate permutations of binary trees}
\usage{
generate_trees(n_layers)
}
\arguments{
\item{n_layers}{Numeric. The number of layers of nodes. Must be greater than 3, should probably be less than 8.}
}
\description{
Create every permutation of binary trees (ignoring chirality) for a given number of layers. These are returned as coordinates defining line segments connecting nodes.
}
