% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tile_rosettes}
\alias{tile_rosettes}
\title{Arrange hexagonal rosettes of binary trees into tessellated layouts.}
\usage{
tile_rosettes(trees, n_rows = 4, n_cols = 4, seed_number = 69)
}
\arguments{
\item{trees}{Data frame. The output from \code{arrange_rosettes()}.}

\item{n_rows}{Numeric. The number of rows to lay the rosettes in. Note that rows are not totally offset from each other in order to tessellate the rosettes. Defaults to \code{4}.}

\item{n_cols}{Numeric. The number of columns in each row to lay the rosettes in. Note that the actual number of rosettes across the results will be twice this. Defaults to \code{4}.}

\item{seed_number}{Numeric. The seed number for the randomization of rosette order. Defaults to \code{69}.}
}
\description{
Creates tessellated arrangement of the provided rosettes.
}
