% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tile_trees}
\alias{tile_trees}
\title{Create tiled rosettes}
\usage{
tile_trees(n_layers, n_rows = 4, n_cols = 4, scale = 1.4, seed_number = 69)
}
\arguments{
\item{n_layers}{Numeric. The number of layers of nodes in each tree. Must be greater than 3, should probably be less than 8.}

\item{n_rows}{Numeric. The number of rows to lay the rosettes in. Note that rows are not totally offset from each other in order to tessellate the rosettes. Defaults to \code{4}.}

\item{n_cols}{Numeric. The number of columns in each row to lay the rosettes in. Note that the actual number of rosettes across the results will be twice this. Defaults to \code{4}.}

\item{scale}{Numeric. The relative scaling to apply to the rosette spacing. Larger values will place the trees farther out from the center. Defaults to \code{1.4}.}

\item{seed_number}{Numeric. The seed number for the randomized processes. Defaults to \code{69}.}
}
\description{
Create a set of binary tree rosettes and tile them
}
