% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{vertices}
\alias{vertices}
\title{Find vertices for a regular polygon}
\usage{
vertices(
  vertex_count,
  radius,
  center_x,
  center_y,
  radian_offset = 0,
  digits = 2
)
}
\arguments{
\item{vertex_count}{Numeric. The number of vertices the polygon has.}

\item{radius}{Numeric. The radius of the circle that the vertices fall along.}

\item{center_x}{Numeric. The x component of the coordinates for the circle's center.}

\item{center_y}{Numeric. The y component of the coordinates for the circle's center.}

\item{radian_offset}{Numeric. The rotational offset for the vertices in radians. If \code{0} then one vertex will be at (cos(0), sin(0)). Defaults to \code{0}.}

\item{digits}{Numeric. The number of decimal places to round the coordinates to. Defaults to \code{2}.}
}
\value{
A data frame with the variables \code{x}, \code{y}, \code{angle}, \code{radius}, \code{cx}, and \code{cy}.
}
\description{
Given a circle defined by radius and center coordinates, find the vertices of a regular polygon inscribed in that circle.
}
